requirements.txt
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
visitor==0.1.3
Werkzeug==0.14.1
